:sparkles: JWT 쿠키 저장 및 리다이렉트 추가

+++ src/main/java/zip/hyeon/snslogintemplate/security/jwt/JwtType.java
@@ -8,9 +8,9 @@ public enum JwtType {
     REFRESH_TOKEN("REFRESH_TOKEN", 60*60*24*7); // 7일
 
     private final String category;
-    private final long expiredTime;
+    private final int expiredTime;
 
-    JwtType(String category, long expiredTime){
+    JwtType(String category, int expiredTime){
         this.category = category;
         this.expiredTime = expiredTime;
     }
